GPSRealtimeHandle/vehicle-location-updater/requirements.txt
pika==1.3.1
python-dotenv==1.0.0 